requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
coverage==5.3
cryptography==3.2.1
cycler==0.10.0
defusedxml==0.7.0rc1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-bootstrap-datepicker-plus
django-environ==0.4.5
django-extensions==3.0.9
django-heroku==0.3.1
djangorestframework==3.11.1
environ==1.0
filelock==3.0.12
flake8==3.8.4
flake8-django==1.1.1
flake8-docstrings==1.5.0
gunicorn==20.0.4
idna==2.10
isort==5.5.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.3.2
mccabe==0.6.1
numpy==1.19.2
oauthlib==3.1.0
Pillow==8.0.0
pipenv==2020.8.13
psycopg2==2.8.5
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pydotplus==2.0.2
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.18.0
six==1.15.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.11
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
